refactor(project): server -> src
