+++ src/command-handler/repository/spending/delete.js
