+++ src/command/checking-account/update-property.js
