+++ src/command/create-monthly-spendings.js
