+++ src/command/spending/delete.js
