+++ src/command/spending/update.js
