+++ src/console-command/accounts.js
