+++ src/console-command/search-reindex.js
