+++ src/model/periodical.js
