+++ src/model/spending.js
