+++ src/repository/periodical.js
